+++ HISTORY.rst
@@ -12,6 +12,7 @@ History
 - Features:
 
   - Provide better message when device is not available.
+  - Backend name stored in host instead of in the AZ (backward incompatible).
 
 0.2.2 (2018-07-24)
 ------------------